pyproject.toml
[project]
name = "ai_reddit_agent"
version = "0.1.0"
description = "AI Reddit News Agent - Automated Reddit discussion analysis and report generation"
authors = [{name = "Developer", email = "dev@example.com"}]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.3",
    "praw>=7.8.1",
    "requests>=2.32.3",
    "beautifulsoup4>=4.13.4",
    "openai>=1.84.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-mock>=3.14.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-mock>=3.14.1",
]
